Moved SwapchainSupport to new file
